backend/requirements.txt
# FastAPI e servidor
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Supabase / HTTP
httpx==0.27.0
python-dotenv==1.0.1
certifi==2024.2.2

# Validação
pydantic==2.6.1
pydantic-settings==2.1.0

# OpenAI (webhook WhatsApp - áudio/texto)
openai==1.12.0

# CORS
starlette==0.36.3
